[Dev] v1.0.0-trt-10.16 (#116)
* v1.0.0-trt-10.16

+++ cookbook/02-API/Layer/RotaryEmbedding/main.py
@@ -1,18 +1,19 @@
-# SPDX-FileCopyrightText: Copyright (c) 1993-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES.
+# All rights reserved.
+#
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-# http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
 import numpy as np
 from tensorrt_cookbook import TRTWrapperV1, case_mark, datatype_cast
@@ -22,9 +23,9 @@ def case_simple():
     b, n, s, h = 1, 2, 4, 8
     max_position = 16
     data = {
-        "input": np.arange(b * n * s * h, dtype=np.float16).reshape(b, n, s, h) / 32,
-        "cos_cache": np.cos(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).astype(np.float16).reshape(max_position, h // 2),
-        "sin_cache": np.sin(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).astype(np.float16).reshape(max_position, h // 2),
+        "input": np.arange(b * n * s * h, dtype=np.float32).reshape(b, n, s, h) / 32,
+        "cos_cache": np.cos(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).reshape(max_position, h // 2),
+        "sin_cache": np.sin(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).reshape(max_position, h // 2),
         "position_ids": np.tile(np.arange(s, dtype=np.int64), (b, 1)),
     }
